msglist: Go edge-to-edge, for a mobile-oriented design
The existing design here was designed for web; the prototype
copied precisely from web's design as of 2022-12.

But in the zulip-mobile RN app, we've long had a design that's
optimized for a narrow mobile screen instead: notably by using the
full available width (up to a maximum, anyway), with no left or right
borders, and consequently less padding too.  Do that here.