android: Bump minSdkVersion to 24, aka Android 7 Nougat
This is the same minimum we have on the zulip-mobile RN app.

We've had it at the Flutter default of 19, aka Android 4.4 KitKat.
But we're now beginning to bump into the need for "multidex" support:
  https://docs.flutter.dev/deployment/android#enabling-multidex-support
at least in debug builds.  The simplest way to do that is to have
minSdkVersion at 21 or higher; which is a good prompt for us to
bump it to the version we were going to want to bump it to anyway.