unreads [nfc]: Use for/in style instead of forEach
Sometimes we get feedback from a lint rule --
  https://dart.dev/tools/linter-rules/avoid_function_literals_in_foreach_calls
-- whose documentation recommends using for loops instead of
forEach.

That rule didn't fire on these instances; not sure why. But anyway
this file has had a mix of for loops and forEach loops, and it seems
nicer to consistently use one or the other.

Just in case there was a performance bug in a Map.forEach
implementation, I tested mark-as-read (by scrolling through the web
app) to see if we might get a free performance benefit there, where
the norm on my device (with a profile build) has been to take about
14ms per event:
  https://github.com/zulip/zulip-flutter/pull/304#issuecomment-1776034014
But, unsurprisingly, it stayed at around 14ms as before.