test: Generate random, increasing message IDs for example data by default
This allows `eg.streamMessage` and `eg.dmMessage` to be used in
contexts where it's important that the message IDs don't collide,
or even that they're increasing, without the test having to explicitly
give specific message IDs when no other fact about them is relevant.

Many call sites that were passing explicit message IDs no longer need
to do so after this change.  We'll clear those out in the next commit.