notif: Navigate to conversation on tapping notification
Still only on Android; iOS support for notifications in general
is coming next.

This also only applies when the app is already running (either in
the foreground or the background) when the notification is opened.
The case where the app isn't already running and is launched from
the notification will be handled in an upcoming commit.

Fixes-partly: #123