api: Cut SendMessageResult.deliverAt, quietly deleted from API
This appears to have been done in commit zulip/zulip@bd2545b0d.
It's fine for it to have gone away -- the new API is better,
and we weren't using this field -- but there's no notice of the
change in the API changelog or the endpoint's own API docs.
That shouldn't happen.  I'll follow up.

Anyway, now that I've sleuthed out the mystery of what happened,
remove this from our bindings to match.