api [nfc]: Document our practice of making constructor arguments explicit
This came up in a code review here:
  https://github.com/zulip/zulip-flutter/pull/361#discussion_r1379374838

and it'd be good to have written down in a more accessible place.

There are some existing parts of the API bindings -- particularly the
parts written in the very first week or two of the prototype, before I
thought through this pattern -- that don't follow this.  I'll fix
those up next.