msglist: Have content pad bottom inset when the compose box isn't shown
This takes advantage of some of StickyHeaderListView's default
behavior (inherited from ListView). From the doc on ListView:

https://api.flutter.dev/flutter/widgets/ListView-class.html
> By default, ListView will automatically pad the list's scrollable
> extremities to avoid partial obstructions indicated by
> MediaQuery's padding. To avoid this behavior, override with a zero
> padding property.

The code changes here are just to arrange for that default behavior
to help us out: we let the real height of the bottom inset propagate
down to the StickyHeaderListView, with the MediaQuery mechanism.