ui: Use 0xfff6f6f6 for ColorScheme.background, following Figma
From logging, this would otherwise be 0xfffefbff.

See the Flutter doc on ColorScheme.background:
  https://api.flutter.dev/flutter/material/ColorScheme/background.html

It simply says:
> A color that typically appears behind scrollable content.

This change is definitely the boldest we've been in perceiving a
pattern in the Figma spec and organizing our code to fit it. We
should cheerfully do something else if it turns out to make it
harder instead of easier to follow the spec.

Prompted by noticing that the surface underneath the scrollable
content in RecentDmConversationsPage wasn't colored according to the
Figma. And that the Figma's design for the Inbox page (#117) had the
same color for the surface underneath *its* scrollable content. This
change fixes the RecentDmConversationsPage and makes it easy and
natural for the Inbox page to get the right color too.