api: Handle surprise removal of email_address from Subscription
This API change was made today in Zulip Server main (at commit
zulip/zulip@432001656), and backported to Zulip Server 7.5, without
discussion of a deprecation period first.  The resulting symptom
is an infinite loading spinner when trying to connect to the server.

A process failure, but fortunately easy to recover from, since this
project is still in alpha.  (And zulip-mobile is unaffected by the
change.)  We don't actually use this field, so just drop it.

Discussion:
  https://chat.zulip.org/#narrow/stream/378-api-design/topic/Removing.20.60email_address.60.20from.20subscription.20objects.3F/near/1683612