store [nfc]: The event-poll loop has-a store, not is-a store
This seems to make more sense conceptually.

For its most immediate practical effect, this means
UpdateMachine.fromInitialSnapshot doesn't care whether
PerAccountStore.fromInitialSnapshot is a generative or a
factory constructor, which will give us the flexibility to
make it the latter.

This change will also probably be helpful for making other
decouplings of the polling and registerNotificationToken logic
from the data structures of PerAccountStore.