gitignore: Ignore detritus from Flutter DevTools
This devtools_options.yaml file seems to get created whenever I
use the Flutter DevTools, the home of the Flutter inspector.

It's pretty vacuous -- just one line reading "extensions:".
Seems like it'd be better for it to not create it.

But in the meantime we can make it not a problem for ourselves,
by just telling Git to ignore it so it doesn't clutter things up.