Try to fix up display names for the app
This is a followup to 1c25b0324, the second commit in this repo's
history: in this round, we try to edit all the user-facing
representations of the name of the app.

Specifically it's "Zulip beta" for the two platforms where we
currently plan to distribute to users in general (so that there's
a risk of confusion if we just said "Zulip"), and just "Zulip" on
the rest.

We don't regularly test on all these platforms, so it's possible I
broke something on, say, Windows.  If so we'll presumably figure
that out if and when we later try to run there.