assets: Add Noto Color Emoji back again, for use on Android
Downloaded from:
  https://github.com/googlefonts/noto-emoji/tree/d79d23e68/fonts

We removed this in a1da95a4e, but we want to add it back now that
we're more confident that it will work on Android. In the time since
a1da95a4e, we learned that COLRv1 fonts just never work on
iOS/macOS; they aren't supported:
  https://github.com/flutter/flutter/issues/134897#issuecomment-1821632095

So, when trying to explain why this font wasn't working in a test on
my iPhone, it seems less necessary to consider a possible unknown
factor that might affect some Android installs too. It was likely
explained just by the fact that the format doesn't have iOS support.

This fixes #435, where the "Licenses" page was broken because we
forgot to remove its reference to the Noto Color Emoji license that
we removed in a1da95a4e.

Fixes: #404
Fixes: #435