msglist: On Android, show reaction chips with Noto Color Emoji font
We should also use this font for Unicode emojis elsewhere, such as
message content, users' names, and so on. But the proper layout of
these chips is especially sensitive to how these emojis get
rendered, and apparently the system emoji font can vary quite a bit
on Android (same link as removed in the TODO):
  https://github.com/zulip/zulip-flutter/pull/410#discussion_r1408403111
So we'd like to converge on one font here (at least for Android) as
soon as possible.