text: Use font-family constants for text built on Typography styles
With plumbing to make kDefaultFamily work, since it's a
variable-weight font.

As it happens, most -- perhaps all -- of the app's text is built on
Typography styles. See the next commit for some thoughts on that.
Since that's the case, I'll mark this commit as fixing these issues:

- #294 Use "Source Sans 3" font for most UI text
- #438 Consistently use "Noto Color Emoji" for emojis on Android

If there's some corner of the app where the two fonts aren't getting
applied, we'll eventually find it, and apply kDefaultFontFamily and
defaultFontFamilyFallback according to their doc comments, to fix.

Fixes: #294
Fixes: #438