text: On Android, render emojis in message content with Noto Color Emoji
In #436, we brought back Noto Color Emoji, but just for reaction
chips. This brings it to message content in the message list, and
all the other places where we've been specifying Source Sans 3.

This leaves some edge cases where we'll still fall back to a system
emoji font, like if there's an emoji in a user's name and you're
looking at the profile screen. (Or in the "sender" area of a message
in the message list, for that matter, checking again just now.)
Those will be fixed later in this series, along with a cleanup that
eliminates the repetition of [kDefaultFontFamily] and
[kDefaultFontFamilyFallback] that we see in this commit.

Fixes-partly: #438