api test: Add smoke test for sendMessage route, covering all params
And in fact this turns up some behavior that seems suspicious!
We have a couple of parameters that we're JSON-encoding, but they're
strings, and I suspect they're not meant to be JSON-encoded.

That's an aspect of the Zulip Server API which is generally not
properly documented: some string parameters get JSON-encoded,
others passed literally.  (In either case the resulting string
is ultimately HTTP-form-encoded to go into the POST request body.)
So the way to learn the answer will be to actually try to use this
feature and see what works.  For now, just leave comments to help
make that debugging easier, by disclaiming the presumption that the
existing code is right.