api: On parse failure, include details in exception
If the server's response is malformed, as with the server issue
that led to #383, the existing logging is unhelpful for working out
what the problem is, even in debug: we learn what request failed,
but not where in the response the misparse was.

When that issue happened, I debugged it with a cruder version of
this change, just augmenting this `catch` block to print the
exception and its stack trace.  But we can do better, using the
magic of Error.throwWithStackTrace.