android: Trust user-supplied CAs, except on known servers
This is copied from the corresponding config in zulip-mobile,
which was added in the following commit:
  https://github.com/zulip/zulip-mobile/commit/85c3a716a7bbf00c936beaf34bc84a0e6bf5c885

Unlike in zulip-mobile, though, we don't need any adjustments
in debug (or profile) builds, because we don't use the
`android:usesCleartextTraffic` attribute which this would clobber.

(React Native debug builds reach out to the development host over
cleartext HTTP in order to support hot updates; but the Flutter
development tools have the host contact the app instead, using the
Dart VM protocol.  It's also not clear that any Dart libraries
respect the `android:usesCleartextTraffic` flag anyway.)

Fixes: #461