content test: Aim taps on links more precisely at the text
By saying `tester.tapAt(find.text('hello'))`, we had been aiming
at the center of the Text widgets, and expecting that to hit the
recognizer we've put on the span, even though the widget is much
wider than the span and the latter doesn't reach the former's center.
Effectively we were relying on the presence of issue #214.

But with an upstream change yesterday:
  https://github.com/flutter/flutter/pull/140621
such a tap no longer hits the span.  That broke these tests.

To fix them, aim the taps near the start of the widget instead.

Fixes: #475