msglist test: Check showSender behavior around date separators
This check in checkInvariants wasn't updated when we added
date separators recently; and the reason nothing broke is because
we didn't have a test to exercise that, oops.

Update the invariant, and extend the relevant test to exercise it.
Now the new test would fail if run with the old checkInvariants.