sticky_header: Handle GrowthDirection.reverse
This resolves the last of the "TODO dir" comments in this library.

We had originally written this library to rely on the assumption
that `constraints.growthDirection` was `GrowthDirection.forward`,
i.e. that the direction in which `constraints.scrollOffset` increases
was the same as `constraints.axisDirection`, in order to keep down
the number of distinct directions one needs to keep track of in
writing and reading (and debugging) the implementation.

But now that this logic is relatively solid within that case --
and now that we have a solid test suite for this library -- we can
go back and identify the handful of places that need to be flipped
around for `GrowthDirection.reverse`, and do so.

This support isn't yet especially *useful*: there are a couple of
unrelated bugs which are triggered when the area allotted to this
sliver isn't the entire viewport, and there isn't a lot of reason
to have a sliver with GrowthDirection.reverse when no other sliver
is potentially occupying part of the viewport.  We'll fix those
separately.