content: Save some work by avoiding `classes.contains` in many cases
It turns out that `classes.contains` instantiates a Set every time
it's called. That's unfortunate.

So, remove some of this work by just checking the class string
(`className`) directly, in the case where we expect just one class.
When we expect two or more classes, we'll want a check that doesn't
enforce a particular order, so perhaps we'll use a regular
expression for those, or something, in later work. See #497.

Fixes-partly: #497