content: Use className regexp for user-mention, instead of class set
This is more efficient but should otherwise behave exactly the same,
except if the className has extraneous whitespace: the previous code
would accept that, and this code will reject as unimplemented.

To help verify that we've gotten this change right, add some more
tests, so that we cover all the different alternatives in this regexp.