content: Use className rather than class set, for CodeBlockSpanNode
This is NFC except in the case where className looks like "  foo  ",
with leading and/or trailing whitespace.  In that case, the previous
code would accept it (if the class name in the middle was one we
recognized), and this code will reject as unimplemented.

For cases where the previous code would reject because `classes.length`
was not 1, this code will reject because className won't be one that
codeBlockSpanTypeFromClassName recognizes.