intellij: Disable spell-check inspection
These `.idea/` directories contain the project configuration and
state of IntelliJ IDEA and its family of IDEs -- in particular
Android Studio, which is what I and Chris use.  There's one for
opening the repo at its root (for working on any of the Dart code)
and another for opening the android/ subfolder (which better
elicits the IDE's support for working on Android-specific code).

Lots of the files here are local state (right down to the size
and location the IDE window last occupied on the screen) that
don't make sense to share.  But mixed in among those is some
project configuration that is best shared; in particular, this
file configuring which "inspections" to run, i.e. what to
highlight in the editor as problems to be fixed.

Start the file off by disabling the spell-check inspection,
because it unavoidably has tons of false positives unless one
maintains a list of all the words it doesn't know.  I've had this
configuration locally since early on; adding to version control
now partly for its own sake, but also as preparation to add
further configuration to this file.