intellij: Disable HTML inspections
The only HTML we have in this project is the example fragments of
Zulip content HTML (as in Zulip message bodies) in a couple of our
test files.  The HTML there is almost all real output from a live
Zulip server; even apart from false-positive complaints, there's no
point critiquing its style because this client's role is to consume
this HTML, not emit it.  So turn all the HTML inspections off.

This clears out a few dozen complaints the various inspections had had
in test/model/content_test.dart in particular, so that it's clean of
issues, just like the rest of our files.  That makes it easier to
navigate to genuine lint issues when they arise while developing
changes, and easier to spot occurrence highlights in the scrollbar
to see where else in the file some identifier appears that one has
one's cursor on.