flutter: Delete broken web platform support
This Flutter app hasn't worked on the web since we started using a
SQLite database, if not sooner.  That's no loss; our actual deployment
targets are (as README.md says) Android and iOS, and even desktop
platforms are supported only for occasional development convenience.

So just cut out the files that make up the Web support, which were
added by `flutter create` from its templates at the beginning and
which we've barely touched since (a total of once, to capitalize
the name "Zulip").