deps [nfc]: Sort lists of dependencies alphabetically
To date we've been adding new dependencies to the end of the
lists.  One consequence is that adding any dependency always
has a merge/rebase conflict with adding any other (unless one
is a dev dependency and the other isn't), because the lines go
in the same place.

That's kind of annoying, so let's start keeping these in
alphabetical order instead.  This way also makes it a bit
easier to find a given dependency.

Also delete a comment the template left about how to configure
lints; we've been successfully doing that when it comes up.