msglist [nfc]: Rename scroll-metrics callbacks more accurately
The main method here was accurately named when first introduced
in 5abeb887e.  But then in bccaead94 its job broadened, so it
should get a more generic name.

(This was basically a mismerge, I think.  The latter commit was
originally written first, so the original draft would have had
its own set of methods, and when I rebased atop the former commit
I apparently didn't do the refactoring that was then called for.)