emoji [nfc]: Move some vertical padding outward
As Greg suggested:
  https://github.com/zulip/zulip-flutter/pull/442#discussion_r1442492101

We drop a comment about giving 1px vertical padding on the label
when it's taller than the emoji (e.g., because the label needs
multiple lines). We still give that padding, and that's a difference
from the web app -- and so arguably worth noting in the code -- but
it's a small point, and clearly an improvement, so it's probably OK
to drop the comment and let the reader use their attention on more
important things.