inbox [nfc]: Cut superfluous StickyHeaderItems on individual conversations
These were never getting consulted, because these aren't entire items
of the enclosing list view; rather the items are the whole sections,
i.e. either a whole stream or all DMs.  Because they weren't doing
anything, they're confusing for understanding how this code works
and making changes to it.

I think this happened because in an early draft of #381, the list-view
items were individual conversations.  That's the arrangement that
would be better for performance/smoothness reasons anyway, but it
won't work until we make some changes to the sticky_header library;
see #389.  When we go to do #389, we can always add these
StickyHeaderItem widgets back.