notif test: In widget tests, use tester.idle instead of `await null`
This calls FakeAsync.flushMicrotasks on the FakeAsync that the
tester is using.  (In the AutomatedTestWidgetsFlutterBinding case,
that is, i.e. the `flutter test` case, which is where the tester
is already using a FakeAsync.)

Seems like it'd be dubious to try to introduce our own FakeAsync
(as with `awaitFakeAsync`) when already in another's zone.  It's
possible it would all work fine, but it seems complicated to
have to think about.