content [nfc]: Use toDiagnosticsNode(name: ...)
This produces slightly nicer output, by indicating the index for
each item.  For example, after introducing an error in one test,
the expected and actual values it prints look like this:

    Actual: <ZulipContent
     └─ListNode
       │ unordered
       │
       ├─item 0: BlockContentNode list
       │ └─ParagraphNode
       │   │ was implicit
       │   │
       │   └─TextNode
       │       "something"
       │
       └─item 1: BlockContentNode list
         └─ParagraphNode
           │ was implicit
           │
           └─TextNode
               "another">

instead of like this:

    Actual: <ZulipContent
     └─ListNode
       │ unordered
       │
       ├─list item
       │ └─ParagraphNode
       │   │ was implicit
       │   │
       │   └─TextNode
       │       "something"
       │
       └─list item
         └─ParagraphNode
           │ was implicit
           │
           └─TextNode
               "another">

This also means less code as we start adding more nodes that have
this kind of structure; it avoids having to make a new subclass of
DiagnosticableTree for each one.