content test: Support localizable UI text in smoke tests
When we support spoilers, soon, we'll want to translate the default
header text ("Spoiler") into the user's language.

It seems safe to assume that Zulip content will always have
`ZulipLocalizations.of` support wherever it appears in the widget
tree. The language will be a client setting with a default value.

There are certain other assumptions that `prepareContentBare` will
probably want to avoid making, though, including:
- that the content is rendered in a particular message
- that the content is rendered in a message at all (#488)
- that the content is rendered in a per-account context (#488)

But we'll think more about that later, when we start using
testContentSmoke in more places.