notif: Add `vm:entry-point` pragma to _onBackgroundMessage
This fixes part of #528, following the suggestion n-92 made here:
  https://github.com/zulip/zulip-flutter/issues/342#issuecomment-1874716442

Even with this change, notifications still don't yet work on Android
in release builds.  But after this change, when the app wasn't in
the foreground, they get farther than before.  Specifically they get
just as far as they get, with or without this change, when the app
was in the foreground.  Details here:
  https://github.com/zulip/zulip-flutter/issues/528#issuecomment-1960646800