tools/git: Fix no_uncommitted_changes to use git-update-index
I was seeing this spuriously report changes when running on macOS,
and fortunately remembered that we'd had a similar problem before
in the various tools in the server-and-web repo.

The fix is in the require_clean_work_tree function there, which in
turn is based on one in Git upstream; see zulip/zulip@73acca76b.