nav [nfc]: Switch to onGenerateRoute instead of home
This has the same effect, but gives us a more general form
of API to work with for more complex effects.

Effectively `onGenerateRoute` is a fallback for when a route name
isn't resolved by `home` or `routes`; and this version has the
same effect as the `home` (and omitted `routes`) we'd been providing.

I find the relevant area of the docs somewhat hard to follow, so to
see this in the implementation: the main effects of both `home` and
`onGenerateRoute` are implemented in [WidgetsApp._onGenerateRoute],
and the one other effect is in [WidgetsApp._usesNavigator].
The [MaterialWidgetRoute] part corresponds to `pageRouteBuilder`,
which is specified by the [MaterialApp] implementation.