notif test: Make tests more robust by providing more data
These tests had been relying on the fact that there wouldn't be any
PerAccountStoreWidget that reached the point of attempting to load
data for these accounts.  But that was true only because, after
opening a notification, the test code waited only for microtasks
(with `await null`), and not for a new frame (like `tester.pump()`).

That makes them brittle, as they'd break on an otherwise innocent
change to pump another frame.  Fortunately it's easy to fix.