ui: Stop setting deprecated ColorScheme.background
We've started getting a deprecation warning from the analyzer:

   info • 'background' is deprecated and shouldn't be used. Use surface instead. This feature
          was deprecated after v3.18.0-0.1.pre • lib/widgets/app.dart:105:9 •
          deprecated_member_use

The advice to use `surface` instead of `background` isn't right for
us, for a few reasons:

1. ColorScheme.fromSeed always produces a ColorScheme with a
   non-null `background`; if you don't pass a `background`, it sets
   one for you, from MaterialDynamicColors. Some code downstream
   (ThemeData's constructor) uses ColorScheme.background
   preferentially over `surface`; I guess that code's behavior helps
   avoid breaking apps that haven't migrated away from `background`.
   Concretely, this means that if we followed the analyzer's
   recommendation to stop passing `background` to
   ColorScheme.fromSeed and pass `surface` instead, then the area
   under scrollable content like the message list would not be our
   chosen color, it would be a default color from
   MaterialDynamicColors.

2. Setting ColorScheme.surface would cause the chosen color to be
   used in at least one part of the UI where it wasn't used before:
   the app bar. We know our app bar isn't finalized, but I take this
   as a sign that `surface` isn't quite the right abstraction to fit
   the Figma's use of this color.

Actually, `background` might already be the wrong abstraction:
currently, it's not only used for the area under scrollable content
like the message list, where the Figma wants it to be used. From
reading code and checking empirically, it's also been getting picked
up and used in two other places: the background of dialogs (as in
our `showErrorDialog`), and our autocomplete popup.

So, use something that seems like a better match for the Figma, and
also works straightforwardly: ThemeData.scaffoldBackgroundColor.
This does mean that dialogs and the autocomplete popup will be a
slightly different shade of near-white than before (that default
`background` courtesy of ColorScheme.fromSeed, the one from
MaterialDynamicColors). That's fine.