store [nfc]: Add direct selfUserId field on PerAccountStore
This makes use sites a bit shorter and more explicit.

We could have made it a getter, but storing the answer directly
here on the PerAccountStore is a space-for-time tradeoff: this
is something we look up frequently, so the extra word of memory
to have a copy directly here seems worth it to save the hash-map
lookup each time.