store [nfc]: Organize the fields and getters a bit more
These have been gradually proliferating, and it's time for another
round of giving more structure to the list.

Also remove a couple of TODO comments that are better expressed
in the tracker; and an admonition about handleEvent that is
probably no longer helpful because there are enough of these fields
that a developer isn't likely to be looking at it when they go to
add another one.