msglist [nfc]: s/_k{FallbackStream,UnsubscribedStreamRecipientHeader}Color/
This name is more parallel with _kDmRecipientHeaderColor.

We could easily have dissolved this constant and just used
AppBarTheme.backgroundColor, which is the same color (and which we
started setting earlier in this series of commits). But it seems
helpful to preserve the constant so it's easy to change. I don't
think the Figma specifies a color for unsubscribed-stream recipient
headers, and I think there's not much reason to prefer this color
over some other color for that.